+++ include/wedit/uid.h
@@ -7,11 +7,11 @@
 using namespace std;
 
 struct UID {   
-    uint8_t site_id;
+    uint32_t site_id;
     uint32_t site_counter;
     vector<uint32_t> global_index;
 
-    UID(uint8_t site_id, uint32_t site_counter, vector<uint32_t> global_index);
+    UID(uint32_t site_id, uint32_t site_counter, vector<uint32_t> global_index);
 };
 
 #endif  // WEDIT_UID_H_
